w3c-reports/12092014/mac/runned-tests/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange-expected.txt

  

PASS test of input.setSelectionRange 
PASS input typeof(input.setSelectionRange)' 
PASS input setSelectionRange return void 
PASS input setSelectionRange(0,1) 
PASS input setSelectionRange(0,input.value.length+1) 
PASS input setSelectionRange(2,2) 
PASS input setSelectionRange(2,1) 
PASS input direction of setSelectionRange(0,1,"backward") 
PASS input direction of setSelectionRange(0,1,"forward") 
PASS input direction of setSelectionRange(0,1,"none") 
PASS input direction of setSelectionRange(0,1,"hoge") 
PASS input direction of setSelectionRange(0,1,"BACKWARD") 
PASS input direction of setSelectionRange(0,1) 
PASS input setSelectionRange("string",1) 
PASS input setSelectionRange(true,1) 
PASS input setSelectionRange([],1) 
PASS input setSelectionRange({},1) 
PASS input setSelectionRange(NaN,1) 
PASS input setSelectionRange(null,1) 
PASS input setSelectionRange(undefined,1) 
PASS test of textarea.setSelectionRange 
PASS textarea typeof(input.setSelectionRange)' 
PASS textarea setSelectionRange return void 
PASS textarea setSelectionRange(0,1) 
PASS textarea setSelectionRange(0,textarea.value.length+1) 
PASS textarea setSelectionRange(2,2) 
PASS textarea setSelectionRange(2,1) 
PASS textarea direction of setSelectionRange(0,1,"backward") 
PASS textarea direction of setSelectionRange(0,1,"forward") 
PASS textarea direction of setSelectionRange(0,1,"none") 
PASS textarea direction of setSelectionRange(0,1,"hoge") 
PASS textarea direction of setSelectionRange(0,1,"BACKWARD") 
PASS textarea direction of setSelectionRange(0,1) 
FAIL textarea setSelectionRange("string",1) assert_equals: element.selectionStart should be 0 expected 1 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:184:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:182:7
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:119:5)
PASS textarea setSelectionRange(true,1) 
FAIL textarea setSelectionRange([],1) assert_equals: element.selectionStart should be 0 expected 1 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:196:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:194:7
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:119:5)
FAIL textarea setSelectionRange({},1) assert_equals: element.selectionStart should be 0 expected 1 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:202:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:200:7
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:119:5)
FAIL textarea setSelectionRange(NaN,1) assert_equals: element.selectionStart should be 0 expected 1 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:208:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:206:7
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:119:5)
FAIL textarea setSelectionRange(null,1) assert_equals: element.selectionStart should be 0 expected 1 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:214:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:212:7
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:119:5)
FAIL textarea setSelectionRange(undefined,1) assert_equals: element.selectionStart should be 0 expected 1 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:220:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:218:7
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/textfieldselection/textfieldselection-setSelectionRange.html:119:5)
